+++ test/BaselineOutput/Common/EntryPoints/core_manifest.json
@@ -21644,6 +21644,7 @@
                   "Auc",
                   "AccuracyMicro",
                   "AccuracyMacro",
+                  "L1",
                   "L2",
                   "F1",
                   "AuPrc",
